Differentiate on message press length

+++ main.py
@@ -3,6 +3,7 @@ import paho.mqtt.client as mqtt
 from pytg import Telegram
 from pytg.receiver import Receiver
 from pytg.sender import Sender
+import json
 import random
 import os
 import requests
@@ -23,12 +24,17 @@ def on_connect(client, userdata, flags, rc):
 def on_message(client, userdata, msg):
     print("I GOT IT")
     print(msg.topic+" "+str(msg.payload))
-    gif = random.choice([x for x in g.search('piglets')])
-    page = requests.get(gif.fixed_height.downsampled.url)
-    with open("/tmp/pig.gif", "wb") as f:
-        f.write(page.content)
-    sender.send_file("Carina", u"/tmp/pig.gif")
-    os.remove("/tmp/pig.gif")
+    age = json.loads(msg.payload)['Age']
+
+    if int(age) > 500:
+        gif = random.choice([x for x in g.search('piglets')])
+        page = requests.get(gif.fixed_height.downsampled.url)
+        with open("/tmp/pig.gif", "wb") as f:
+            f.write(page.content)
+        sender.send_file("Carina", u"/tmp/pig.gif")
+        os.remove("/tmp/pig.gif")
+    else
+        sender.send_msg("Carina", u"🐕")
 
 client = mqtt.Client()
 client.on_connect = on_connect
